backend/llm-email/requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-socketio==5.11.2

# Data Models & Validation
pydantic==2.8.2

# Machine Learning / NLP
torch==2.4.0+cpu      # CPU-only build (Windows safe)
torchvision==0.19.0+cpu
torchaudio==2.4.0+cpu
transformers==4.43.4
joblib==1.4.2
scikit-learn==1.5.1
numpy==1.26.4

# Google APIs (for Gmail integration etc.)
google-api-python-client==2.147.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Utils
requests==2.32.3
httpx==0.27.0
